Update README.md
This project analyzes the police shooting data between different regions in the United States and stores it into two different data structures. One of the data structures is a Max Binary heap while the other is a Splay tree. After completing the project we concluded that the Max Binary heap was more efficent in finding regions with the most shootings.